FIX: reorganization of alphafold format

+++ structuremap/processing.py
@@ -152,7 +152,6 @@ def format_alphafold_data(directory: str,
                                    'x_coord':structure['_atom_site.Cartn_x'],
                                    'y_coord':structure['_atom_site.Cartn_y'],
                                    'z_coord':structure['_atom_site.Cartn_z']})
-                df["protein_number"] = df["protein_number"].astype(np.int32)
 
                 df = df[df.atom_id.isin(['CA','CB','C','N'])].reset_index(drop=True)
                 df = df.pivot(index=['protein_id','protein_number','AA','position','quality'], columns="atom_id")
@@ -200,7 +199,7 @@ def format_alphafold_data(directory: str,
     structure_types = list(alphafold_annotation.structure_group.unique())
     str_oh = pd.get_dummies(alphafold_annotation['structure_group'], dtype='int64')
     alphafold_annotation = alphafold_annotation.join(str_oh)
-
+    alphafold_annotation["protein_number"] = alphafold_annotation["protein_number"].astype(np.int32)
     return(alphafold_annotation)
 
 @numba.njit
